add the ability for modular tokenization
From now on, the collection of regexes is saved in a dictionary on disk.
The dictionary is generated from a script, but any user can add his own
regexes directly in the text file.

Moreover, a new tokenizer class has been created, which allows for the
creation of arbitrary tokenization pipelines.